Move SearchableSnapshotsUtils from x-pack core to plugin (#75299)
